refactor(plist): move tests to `src` folder
